requirements.txt
fastapi
uvicorn[standard]
yclients-api
httpx
pydantic-settings
email-validator